#3 Сделать ручку GET /status (Исправлен тест для InfoController)
