+++ backend/src/test/java/ru/eda/tech/TechApplicationTest.java
@@ -1,0 +1,13 @@
+package ru.eda.tech;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class TechApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
