+++ backend/src/test/java/ru/eda/tech/controller/InfoControllerTest.java
@@ -1,0 +1,27 @@
+package ru.eda.tech.controller;
+
+import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
+import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.print;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.content;
+import static org.hamcrest.Matchers.equalTo;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
+import org.springframework.test.web.servlet.MockMvc;
+
+@WebMvcTest(InfoController.class)
+class InfoControllerTest {
+
+    @Autowired
+    private MockMvc mockMvc;
+
+    @Test
+    void getStatus() throws Exception {
+        this.mockMvc.perform(get("/status"))
+                .andDo(print())
+                .andExpect(status().isOk())
+                .andExpect(content().string(equalTo("")));
+    }
+}
